requirements.txt
cffi==1.15.0
fastcore==1.3.27
ghapi==0.1.19
loguru==0.6.0
packaging==21.3
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.7
python-dotenv==0.19.2
